+++ mindsdb/api/executor/sql_query/steps/map_reduce_step.py
@@ -1,4 +1,3 @@
-import os
 import copy
 
 from mindsdb_sql_parser.ast import (
@@ -15,8 +14,7 @@ from mindsdb.api.executor.planner.steps import (
 
 from mindsdb.api.executor.sql_query.result_set import ResultSet
 from mindsdb.api.executor.exceptions import LogicError
-from mindsdb.utilities.config import Config
-from mindsdb.utilities.context_executor import execute_in_threads
+from mindsdb.utilities.context_executor import process_dataframe_in_partitions
 
 from .base import BaseStepCall
 
@@ -88,43 +86,12 @@ class MapReduceStepCall(BaseStepCall):
 
         df = input_data.get_raw_df()
 
-        # tasks
-        def split_data_f(df):
-            chunk = 0
-            while chunk * partition < len(df):
-                # create results with partition
-                df1 = df.iloc[chunk * partition: (chunk + 1) * partition]
-                chunk += 1
-                yield df1, substeps, input_idx, input_columns
+        def callback(chunk):
+            return self._exec_partition(chunk, substeps, input_idx, input_columns)
 
-        tasks = split_data_f(df)
-
-        # workers count
-        is_cloud = Config().get('cloud', False)
-        if is_cloud:
-            max_threads = int(os.getenv('MAX_QUERY_PARTITIONS', 10))
-        else:
-            max_threads = os.cpu_count() - 2
-
-        # don't exceed chunk_count
-        chunk_count = int(len(df) / partition)
-        max_threads = min(max_threads, chunk_count)
-
-        if max_threads < 1:
-            max_threads = 1
-
-        if max_threads == 1:
-            # don't spawn threads
-
-            for task in tasks:
-                sub_data = self._exec_partition(*task)
-                if sub_data:
-                    data = join_query_data(data, sub_data)
-
-        else:
-            for sub_data in execute_in_threads(self._exec_partition, tasks, thread_count=max_threads):
-                if sub_data:
-                    data = join_query_data(data, sub_data)
+        for result in process_dataframe_in_partitions(df, callback, partition):
+            if result:
+                data = join_query_data(data, result)
 
         return data
